requirements.txt
python-docx==1.1.0
pandas==2.2.3
openpyxl==3.1.2
pdf2image==1.17.0
pdfplumber==0.11.5
Pillow==11.1.0
protobuf==5.29.3
pydantic==2.10.6
PyPDF2==3.0.1
pytest==8.3.4
python-dotenv==1.0.1
PyYAML==6.0.2b
PyYAML==6.0.2
setuptools==75.8.0
watchdog==6.0.0
